uint8_t m_frameIndex.

+++ NonuniformBlur/NonuniformBlur.h
@@ -60,7 +60,7 @@ private:
 	float		m_sigma;
 
 	// Synchronization objects.
-	uint32_t	m_frameIndex;
+	uint8_t		m_frameIndex;
 	HANDLE		m_fenceEvent;
 	XUSG::Fence	m_fence;
 	uint64_t	m_fenceValues[FrameCount];
